Squashed commit of the following:
commit b3dcdc96d6e5834fcb32c90beedc5b179c941d54
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Wed Mar 15 17:10:09 2023 +0300

    Currency migration fixes (#69)

commit dbf6a3ad393d55bae9bbf39cfc2236a8ff674be7
Author: Leonardo Hildt <108893281+leonardohildt@users.noreply.github.com>
Date:   Fri Mar 10 13:55:04 2023 -0300

    Fix to handle the exception (#63)

    This will avoid so many exceptions in the logs

commit 149783fcc1fe5671813fab073b9885b7b91fe61a
Author: Leonardo Hildt <108893281+leonardohildt@users.noreply.github.com>
Date:   Fri Mar 10 13:53:52 2023 -0300

    Fix build error (#65)

commit fd53a56b81c1613133e7d316f26ed713012a8096
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Fri Mar 10 19:32:53 2023 +0300

    ABA-18 and ABA-75 (#64)

    * ABA-74 Vat Calculation for Interest

    * Improve Calculations

    * Update 0031_create_vat_rate_to_client.xml

    * Fix transactions breakdown

    * ABA-18 and ABA-75

commit ce729321290d2c96d95696c583ce7f5bc0a81b96
Merge: 9f155b4df 684b402f7
Author: akapusFITER <108024337+akapusFITER@users.noreply.github.com>
Date:   Thu Mar 9 19:19:58 2023 -0300

    Merge pull request #62 from FITER1/fix/ABA-20-v1

    ABA-20 Adjustment on installments for VAT

commit 684b402f7ff60913d497d2f923013aa25556b226
Author: Leonardo Hildt <leonardo@fiter.io>
Date:   Thu Mar 9 19:16:57 2023 -0300

    ABA-20 Adjustment on installments for VAT

commit 9f155b4dfa02fba625624b7a3ce2e76e9d2863bc
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Fri Mar 10 00:48:35 2023 +0300

    ABA-74 (#61)

    * ABA-74 Vat Calculation for Interest

    * Improve Calculations

    * Update 0031_create_vat_rate_to_client.xml

    * Fix transactions breakdown

commit 0b97f4b4d6924f24b8b4e24cf3f99c86e3c14764
Author: Leonardo Hildt <108893281+leonardohildt@users.noreply.github.com>
Date:   Wed Mar 8 14:25:38 2023 -0300

    ABA-20 1er cuota irregular adaptada a los días de duración (#60)

commit f970076ecce326a8e3d242c5144aa9c2fdeb16de
Author: hamidoke <110378533+hamidoke@users.noreply.github.com>
Date:   Wed Mar 8 16:25:08 2023 +0100

    Update build.yml


commit 76f38c27b2522d54a4f1cb69af7efcca01eceb6b
Author: hamidoke <110378533+hamidoke@users.noreply.github.com>
Date:   Wed Mar 8 15:58:20 2023 +0100

    Update build.yml


commit c761119929d91baa01bb0d8b70f987cbcdb37fbc
Author: hamidoke <110378533+hamidoke@users.noreply.github.com>
Date:   Wed Mar 8 15:49:45 2023 +0100

    Update build.yml


commit 3b06dc4fa29bc9f95b404679d30b02c583914227
Author: hamidoke <110378533+hamidoke@users.noreply.github.com>
Date:   Wed Mar 8 15:27:52 2023 +0100

    Update build.yml


commit c814a49a90ebd763e98c9f34f2df0ff5137374ba
Author: hamidoke <110378533+hamidoke@users.noreply.github.com>
Date:   Wed Mar 8 15:25:10 2023 +0100

    Update build.yml


commit 221879576b9cd44192d6626686f779a7be112fea
Merge: 9b9e03fe3 a7e1127dc
Author: akapusFITER <108024337+akapusFITER@users.noreply.github.com>
Date:   Wed Mar 1 09:03:42 2023 -0300

    Merge pull request #59 from FITER1/feature/ABA-61

    Feature/ABA-61

commit a7e1127dc5d47d0725bef62103aa4285c4d38199
Author: Leonardo Hildt <leonardo@fiter.io>
Date:   Tue Feb 28 10:41:57 2023 -0300

    ABA-61 Agregar el CAT con IVA y Sin IVA al préstamo

commit 9b9e03fe39e678626209e4193ccc54fd8d41ceb9
Merge: 43f74fb2a 2b40d99ef
Author: akapusFITER <108024337+akapusFITER@users.noreply.github.com>
Date:   Thu Feb 23 12:09:47 2023 -0300

    Merge pull request #57 from FITER1/ABA-54

    ABA-54

commit 2b40d99ef62d7f366a4719d2ae725fc5633b5a9b
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Thu Feb 23 18:03:04 2023 +0300

    A bit more changes

commit 43f74fb2ace028bd6b67ef22839f0692436f2290
Merge: d1aa44015 93a648b71
Author: akapusFITER <108024337+akapusFITER@users.noreply.github.com>
Date:   Wed Feb 22 10:26:13 2023 -0300

    Merge pull request #55 from FITER1/feature/ABA-46

    ABA-46 Agregar el tab de Situación Crediticia

commit ab99318612ee47d2226a66da018c176f11ab0eec
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Wed Feb 22 14:59:01 2023 +0300

    ABA-54

commit 93a648b711033b8589de2961c2055fcd60267ccd
Author: Leonardo Hildt <leonardo@fiter.io>
Date:   Tue Feb 21 16:01:20 2023 -0300

    ABA-46 Agregar el tab de Situación Crediticia

commit d1aa44015c9868be0a7a93035ad3cb3a7741d62d
Merge: abc548464 95df7188d
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Fri Feb 17 19:56:50 2023 +0300

    Merge pull request #47 from FITER1/ABA-59

    ABA-59

commit 95df7188dea7d3f6a23d247d81ee0b8a48b624b7
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Fri Feb 17 19:21:02 2023 +0300

    ABA-59

commit abc548464bbc140de7d0d4346a7052fab11139cb
Merge: d1ff23bd6 05d7e3cac
Author: akapusFITER <108024337+akapusFITER@users.noreply.github.com>
Date:   Fri Feb 17 15:42:20 2023 +0200

    Merge pull request #46 from FITER1/feature/aba-45

    ABA-45 Agregar el link para el tab de documentos

commit 05d7e3cacb92fbfd5f1bf155748cae1e583a4d5e
Author: Leonardo Hildt <leonardo@fiter.io>
Date:   Fri Feb 17 10:35:03 2023 -0300

    ABA-45 Agregar el link para el tab de documentos

commit d1ff23bd62570f23e4a70f91801a34d2d8e4cc70
Merge: 6dfa01c19 42dbabfe7
Author: akapusFITER <108024337+akapusFITER@users.noreply.github.com>
Date:   Fri Feb 17 01:23:27 2023 +0200

    Merge pull request #45 from FITER1/ABA-52_53

    ABA-52 and 53

commit 42dbabfe78205addd0c4a777fa40d6b12d275d62
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Fri Feb 17 02:06:32 2023 +0300

    Merge conflict fixes

commit 211f670686b410e40ae6547919aefdeb3872cb97
Merge: 404c1e603 6dfa01c19
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Fri Feb 17 01:54:30 2023 +0300

    Merge branch 'fiter/albo/dev' into ABA-52_53

    # Conflicts:
    #	fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/data/LoanAccountData.java
    #	fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/domain/Loan.java
    #	fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/service/LoanReadPlatformServiceImpl.java

commit 404c1e603ae1aceb6d88d1042878003aed5383b4
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Fri Feb 17 01:25:21 2023 +0300

    ABA-52 and 53

commit 6dfa01c192f3094bd234ca7d6f3f4968b35f0747
Merge: 8f113bc21 531b89d42
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Thu Feb 16 22:17:41 2023 +0300

    Merge pull request #44 from FITER1/feature/ABA-40

    ABA-40 CAT calculation

commit 531b89d428f04ae452b0f07e3ec02649a8c8f501
Author: Leonardo Hildt <leonardo@fiter.io>
Date:   Wed Feb 15 19:33:31 2023 -0300

    ABA-40 CAT calculation

commit 8f113bc21d4093fe89f805a132ffe63247946e13
Merge: d6e46f641 db1fdedc1
Author: akapusFITER <108024337+akapusFITER@users.noreply.github.com>
Date:   Wed Feb 15 14:48:11 2023 +0200

    Merge pull request #42 from FITER1/ABA-51

    ABA-51

commit db1fdedc1fbe19d8bbbba30162dc1388d089adb6
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Wed Feb 15 09:31:09 2023 +0300

    ABA-51

commit d6e46f641fe16b04a10b5c904acac75979e6fe66
Merge: eead5bb42 d11efe5fe
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Wed Feb 15 01:43:29 2023 +0300

    Merge pull request #41 from FITER1/ABA-50

    ABA-50 create Vat Rate tables

commit d11efe5fecd4737f99d0811a84df8ed85dc81b89
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Tue Feb 14 13:17:16 2023 +0300

    Update 0031_create_vat_rate_to_client.xml

commit 8d48eae0f47e4c72a083ca336b8a7aa660543712
Merge: 843fba68e eead5bb42
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Tue Feb 14 11:39:56 2023 +0300

    Merge branch 'fiter/albo/dev' into ABA-50


commit 843fba68e3cd92a1e8dd2415a1adefaad081d98d
Author: Magezi Arthur <artmagezi@gmail.com>
Date:   Tue Feb 14 11:36:09 2023 +0300

    ABA-50 create Vat Rate tables

+++ fineract-provider/src/main/java/org/apache/fineract/infrastructure/documentmanagement/service/DocumentWritePlatformServiceJpaRepositoryImpl.java
@@ -71,11 +71,16 @@ public class DocumentWritePlatformServiceJpaRepositoryImpl implements DocumentWr
 
             final ContentRepository contentRepository = this.contentRepositoryFactory.getRepository();
 
-            final String fileLocation = contentRepository.saveFile(inputStream, documentCommand);
+            String fileName = "";
+            String fileLocation = "";
+            if (documentCommand.getFileName() != null && inputStream != null) {
+                fileName = documentCommand.getFileName();
+                fileLocation = contentRepository.saveFile(inputStream, documentCommand);
+            }
 
             final Document document = Document.createNew(documentCommand.getParentEntityType(), documentCommand.getParentEntityId(),
-                    documentCommand.getName(), documentCommand.getFileName(), documentCommand.getSize(), documentCommand.getType(),
-                    documentCommand.getDescription(), fileLocation, contentRepository.getStorageType());
+                    documentCommand.getName(), fileName, documentCommand.getSize(), documentCommand.getType(),
+                    documentCommand.getDescription(), fileLocation, contentRepository.getStorageType(), documentCommand.getDocumentLink());
 
             this.documentRepository.saveAndFlush(document);
 
@@ -93,7 +98,7 @@ public class DocumentWritePlatformServiceJpaRepositoryImpl implements DocumentWr
             final String mimeType, final String name, final String description, final String fileName) {
 
         final DocumentCommand documentCommand = new DocumentCommand(null, null, entityType, entityId, name, fileName, fileSize, mimeType,
-                description, null);
+                description, null, null);
 
         final Long documentId = createDocument(documentCommand, inputStream);
 
@@ -156,8 +161,10 @@ public class DocumentWritePlatformServiceJpaRepositoryImpl implements DocumentWr
                         documentCommand.getId()));
         this.documentRepository.delete(document);
 
-        final ContentRepository contentRepository = this.contentRepositoryFactory.getRepository(document.storageType());
-        contentRepository.deleteFile(document.getLocation());
+        if (document.getLocation() != null) {
+            final ContentRepository contentRepository = this.contentRepositoryFactory.getRepository(document.storageType());
+            contentRepository.deleteFile(document.getLocation());
+        }
         return new CommandProcessingResult(document.getId());
     }
